requirements.txt
bidict
gym
numpy
pandas
pygame
PyYAML
scipy
ipykernel
matplotlib